nh-search: extract terminal rendering
