+++ crates/nh-search/src/lib.rs
@@ -3,5 +3,6 @@ pub mod args;
 mod backend;
 mod channel;
 mod query;
+mod render;
 mod search;
 mod types;
